requirements.txt
absl-py==2.3.0
astunparse==1.6.3
bioimageio.core==0.5.11
bioimageio.spec==0.4.9.post5
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
contourpy==1.1.1
csbdeep==0.8.1
cycler==0.12.1
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
future==1.0.0
gast==0.4.0
google-auth==2.40.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.41.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-jsonschema==0.13.0
marshmallow-union==0.1.15.post1
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
nodejs==0.1.1
numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.19.2
opt_einsum==3.4.0
optional-django==0.1.0
packaging==25.0
pandas==2.0.3
pillow==10.4.0
protobuf==4.25.8
pyaml==25.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.8
scikit-image==0.21.0
scikit-learn==1.3.2
scikit-optimize==0.10.2
scipy==1.10.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
stardist==0.9.1
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
triton==3.0.0
typer==0.16.0
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.2
xarray==2023.1.0
zipp==3.20.2
